fixing chroot logging
more packages ignored...

package updates, and it keeps dying on this strange point in the build.

think i figured out the bug with the build..

oh, why the hell not.

i really hope this works.

dang it, 'root' is not a variable.

okay. think i figured it out. nspawn doesn't like pipes.